cherry-pick(#24213): Revert "fix: do not collide with other tests when test names have special chars (#23414)" (#24217)
This PR cherry-picks the following commits:

- 57cca1d96e34ab9242570abc2464302d80cde504
